GameObject class

+++ Based_Engine/Based_Engine/Application.cpp
@@ -48,7 +48,7 @@ bool Application::Init()
 // ---------------------------------------------
 void Application::PrepareUpdate()
 {
-	dt = (float)ms_timer.Read() / 1000.0f;
+	dt = ms_timer.Read() / 1000.0f;
 	ms_timer.Start();
 }
